🎨 feat(colors): totally refactor color usage to make it possible to re-theme the app
Prior to this commit, a lot of colors where hardoded inside the "style"
section of the .vue files. The color variables available for
configuration in the variables.scss file and their usage was also
blurry.

This commit totally overhauls that, defining color variables by purpose
and using them everywhere accordingly, effectively removing any
hardcoded colors and allowing to re-theme the schedule more easily.